(maint) Update the acceptance tests
Prior to this commit, the acceptance tests for mailalias were broken up
in a way that was not logical. The tests had been broken up simply
following the 'steps' from the previous iteration. This new structure
did not work, since later steps were dependent on earlier steps. The
tests could not be run in isolation. This commit changes the test set up
so that we can more easily run tests in isolation and in random order.