Use reusable static code analysis
This commit updates the static code analysis GitHub Actions workflow
from using a copy in this repository to a resuable copy in the Phoenix
team centralized GitHub Actions repository.

In addition, this removes the commits Rake task as it is no longer used
in the reusable version of this workflow.